auth_service/requirement.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
pydantic 
python-jose[cryptography] 
passlib[bcrypt] 
bcrypt 
python-dotenv 
redis 
requests 
pytest 
responses 
alembic
grpcio grpcio-tools